99-integrations/openclaw-imports/mcp-client/requirements.txt
mcp>=1.0.0
pyyaml>=6.0
click>=8.0.0
anyio>=4.0.0